mediconnect_backend/requirements.txt
apscheduler
sqlalchemy
pydantic-settings
asyncpg
google-generativeai
python-dotenv
fastapi-limiter